requirements-hotkey.txt
pynput==1.8.1
pystray>=0.19.5
Pillow>=10.0.0
webrtcvad-wheels>=2.0.11
python-dotenv>=1.0.0
